faust_worker/requirements.txt
confluent-kafka>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.3
# Uncomment if vector search is needed
# faiss-cpu>=1.7.4